// src/shared/api/generated/swaggerTypes.ts
export interface UserResponse {
  id: number;
  identification: string | number;
  fullname: string;
  first_name?: string;
  last_name?: string;
  email: string;
  phone?: string;
  address?: string;
  role: 'Administrador' | 'Instructor' | 'Aprendiz';
  status: boolean | string;
  is_active?: boolean;
  created_at?: string;
  updated_at?: string;
}